update ops register
